+++ PowerPointLabs/PowerPointLabs/ImageSearch/AdjustImageWindow.xaml
@@ -1,7 +1,7 @@
 ﻿<controls:MetroWindow 
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    x:Class="PowerPointLabs.ImageSearch.Crop.CropWindow"
+    x:Class="PowerPointLabs.ImageSearch.AdjustImageWindow"
     xmlns:controls="clr-namespace:MahApps.Metro.Controls;assembly=MahApps.Metro"
     Title="Adjust Image Offset"
     WindowStartupLocation="CenterScreen"
