+++ PowerPointLabs/PowerPointLabs/ImageSearch/SearchOptionsPane.Obsolete.xaml.cs
