+++ PowerPointLabs/PowerPointLabs/ImageSearch/StyleOptionsPane.Obsolete.xaml
