+++ PowerPointLabs/PowerPointLabs/ImageSearch/StyleOptionsPane.Obsolete.xaml.cs
